+++ src/content/inotify/inotify_handler.cc
@@ -33,6 +33,7 @@
 #include "content/content.h"
 #include "content/inotify/autoscan_inotify.h"
 #include "content/inotify/directory_watch.h"
+#include "content/inotify/inotify_types.h"
 #include "content/inotify/watch.h"
 #include "database/database.h"
 #include "util/logger.h"
@@ -54,82 +55,13 @@
 #define AUTOSCAN_IS_WRITTEN(mask) ((mask) & (IN_CLOSE_WRITE | IN_MOVE_SELF | IN_ATTRIB))
 #define AUTOSCAN_IS_IGNORED(mask) ((mask) & (IN_IGNORED))
 
-#ifndef IN_MASK_CREATE
-#define IN_MASK_CREATE 0x10000000
-#endif
-
-static constexpr std::array<std::pair<std::string_view, InotifyFlags>, 22> inotifyFlags {
-    /* the following are legal, implemented events that user-space can watch for */
-    std::pair("ACCESS", IN_ACCESS), /* File was accessed */
-    std::pair("MODIFY", IN_MODIFY), /* File was modified */
-    std::pair("ATTRIB", IN_ATTRIB), /* Metadata changed */
-    std::pair("CLOSE_WRITE", IN_CLOSE_WRITE), /* Writable file was closed */
-    std::pair("CLOSE_NOWRITE", IN_CLOSE_NOWRITE), /* Unwritable file closed */
-    std::pair("OPEN", IN_OPEN), /* File was opened */
-    std::pair("MOVED_FROM", IN_MOVED_FROM), /* File was moved from X */
-    std::pair("MOVED_TO", IN_MOVED_TO), /* File was moved to Y */
-    std::pair("CREATE", IN_CREATE), /* Subfile was created */
-    std::pair("DELETE", IN_DELETE), /* Subfile was deleted */
-    std::pair("DELETE_SELF", IN_DELETE_SELF), /* Self was deleted */
-    std::pair("MOVE_SELF", IN_MOVE_SELF), /* Self was moved; Generates an IN_MOVED_FROM event for dir1, an IN_MOVED_TO event for dir2 */
-    /* the following are legal events.  they are sent as needed to any watch */
-    std::pair("UNMOUNT", IN_UNMOUNT), /* Backing fs was unmounted */
-    std::pair("Q_OVERFLOW", IN_Q_OVERFLOW), /* Event queued overflowed */
-    std::pair("IGNORED", IN_IGNORED), /* File was ignored */
-    /* special flags */
-    std::pair("ONLYDIR", IN_ONLYDIR), /* only watch the path if it is a directory */
-    std::pair("DONT_FOLLOW", IN_DONT_FOLLOW), /* don't follow a sym link */
-    std::pair("EXCL_UNLINK", IN_EXCL_UNLINK), /* exclude events on unlinked objects */
-    std::pair("MASK_CREATE", IN_MASK_CREATE), /* only create watches */
-    std::pair("MASK_ADD", IN_MASK_ADD), /* add to the mask of an already existing watch */
-    std::pair("ISDIR", IN_ISDIR), /* event occurred against dir */
-    std::pair("ONESHOT", IN_ONESHOT), /* only send event once */
-};
-
 InotifyHandler::InotifyHandler(AutoscanInotify* ai, struct inotify_event* event, InotifyFlags maskedEvent)
     : ai(ai)
     , wd(event->wd)
     , mask(maskedEvent)
     , name(event->len > 0 ? event->name : "") // The name field is only present when an event is returned for a file inside a watched directory
 {
-    log_debug("inotify event: {} mask={} name={}", wd, mapFlags(mask), name);
-}
-
-InotifyFlags InotifyHandler::remapFlag(const std::string& flag)
-{
-    for (auto [qLabel, qFlag] : inotifyFlags) {
-        if (flag == qLabel) {
-            return qFlag;
-        }
-    }
-    return stoulString(flag, 0, 0);
-}
-
-InotifyFlags InotifyHandler::makeFlags(const std::string& optValue)
-{
-    std::vector<std::string> flagsVector = splitString(optValue, '|', false);
-    return std::accumulate(flagsVector.begin(), flagsVector.end(), 0, [](auto flg, auto&& i) { return flg | InotifyHandler::remapFlag(trimString(i)); });
-}
-
-std::string InotifyHandler::mapFlags(InotifyFlags flags)
-{
-    if (!flags)
-        return "None";
-
-    std::vector<std::string> myFlags;
-
-    for (auto [qLabel, qFlag] : inotifyFlags) {
-        if (flags & qFlag) {
-            myFlags.emplace_back(qLabel);
-            flags &= ~qFlag;
-        }
-    }
-
-    if (flags) {
-        myFlags.push_back(fmt::format("{:#04x}", flags));
-    }
-
-    return fmt::format("{}", fmt::join(myFlags, " | "));
+    log_debug("inotify event: {} mask={} name={}", wd, InotifyUtil::mapFlags(mask), name);
 }
 
 fs::path InotifyHandler::getPath(const std::shared_ptr<DirectoryWatch>& wdObj)
@@ -138,9 +70,9 @@ fs::path InotifyHandler::getPath(const std::shared_ptr<DirectoryWatch>& wdObj)
     // file is not gone
     if (!AUTOSCAN_IS_GONE(mask) && !name.empty()) {
         path /= name;
-        log_debug("!AUTOSCAN_IS_GONE({}): {}", mapFlags(mask), path.c_str());
+        log_debug("!AUTOSCAN_IS_GONE({}): {}", InotifyUtil::mapFlags(mask), path.c_str());
     } else {
-        log_debug("processing {} event {}", path.c_str(), mapFlags(mask));
+        log_debug("processing {} event {}", path.c_str(), InotifyUtil::mapFlags(mask));
     }
     return path;
 }
@@ -161,7 +93,7 @@ std::pair<bool, std::shared_ptr<AutoscanDirectory>> InotifyHandler::getAutoscanD
     if (!watchAs || !adir) {
         log_debug("autoscan not found in watches? ({}, watchAs:{}, adir:{}, {})", wd, watchAs != nullptr, adir != nullptr, path.c_str());
     } else if (ec && !AUTOSCAN_WAS_REMOVED(mask)) {
-        log_error("Failed to read {} for event {}: {}", path.c_str(), mapFlags(mask), ec.message());
+        log_error("Failed to read {} for event {}: {}", path.c_str(), InotifyUtil::mapFlags(mask), ec.message());
     }
     log_debug("autoscan for watch (isDir:{}, adir:{}, {})", isDir, adir != nullptr, path.c_str());
     return { isDir, adir };