+++ test/mock/config_mock.h
@@ -44,6 +44,7 @@ public:
     }
     std::uint32_t getUIntOption(ConfigVal option) const override { return 0; }
     std::int64_t getLongOption(ConfigVal option) const override { return 0; }
+    unsigned long long getULongOption(ConfigVal option) const override { return 0; }
     std::shared_ptr<ConfigOption> getConfigOption(ConfigVal option) const override { return {}; }
     bool getBoolOption(ConfigVal option) const override
     {
@@ -142,6 +143,7 @@ public:
     void setOrigValue(const std::string& item, std::int32_t value) override { }
     void setOrigValue(const std::string& item, std::uint32_t value) override { }
     void setOrigValue(const std::string& item, std::int64_t value) override { }
+    void setOrigValue(const std::string& item, unsigned long long value) override { }
     bool hasOrigValue(const std::string& item) const override { return false; }
     MOCK_METHOD(std::shared_ptr<TranscodingProfileList>, getTranscodingProfileListOption, (ConfigVal option), (const override));
     MOCK_METHOD(std::shared_ptr<DynamicContentList>, getDynamicContentListOption, (ConfigVal option), (const override));