+++ ChangeLog.md
@@ -2,7 +2,9 @@
 
 ### HEAD
 
+- Bump tmp from 0.2.3 to 0.2.4 in /gerbera-web
 - Fix MySql config load
+- Inotify for scripts: reload without server restart
 
 ### v2.6.0